refactor: extract editor opening logic into a reusable trait
The `openInEditor` method was duplicated across the `Llm\Commit` and `Llm\Pr` commands. This change extracts the shared logic into a new `HandlesLlmOutput` trait located at `app/Console/Commands/Llm/Concerns/HandlesLlmOutput.php`.

Both `Llm\Commit` and `Llm\Pr` commands now utilize this trait, reducing code duplication and improving maintainability. Consistent error handling for the editor process is also consolidated within the trait.
